Inpainting/requirements.txt
aiofiles==23.2.1
aiohttp==3.9.5
aiosignal==1.3.1
altair==5.3.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.4.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.2.1
async-timeout==4.0.3
asyncer==0.0.7
attrs==23.1.0
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.0.0
blinker==1.7.0
cachetools==5.3.3
certifi==2023.11.17
cffi==1.15.1
charset-normalizer==3.3.2
click==8.1.7
coloredlogs==15.0.1
comm==0.1.3
contourpy==1.2.0
cupy-cuda112==10.6.0
cupy-cuda11x==12.2.0
cycler==0.12.1
Cython==3.0.6
cython-bbox==0.1.5
dataclasses-json==0.6.3
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
diffusers==0.29.2
dnspython==2.6.1
email_validator==2.2.0
exceptiongroup==1.1.1
executing==1.2.0
fastapi==0.111.0
fastapi-cli==0.0.4
fastjsonschema==2.17.1
fastrlock==0.8.2
ffmpy==0.3.2
filelock==3.13.1
filetype==1.2.0
flatbuffers==24.3.25
fonttools==4.45.1
frozenlist==1.4.1
fsspec==2023.10.0
gitdb==4.0.11
GitPython==3.1.43
gradio==4.37.2
gradio_client==1.0.2
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.23.4
humanfriendly==10.0
idna==3.6
imageio==2.34.2
importlib-metadata==6.8.0
importlib_resources==6.4.0
ipykernel==6.23.3
ipython-genutils==0.2.0
jedi==0.18.2
Jinja2==3.1.2
jsonschema==4.20.0
jsonschema-specifications==2023.11.1
jupyter-events==0.6.3
jupyter_client==8.3.0
jupyter_core==5.3.1
jupyter_server==2.7.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
kiwisolver==1.4.5
lap==0.4.0
lazy_loader==0.4
llvmlite==0.43.0
loguru==0.7.2
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib==3.8.2
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.1
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
nbclassic==1.0.0
nbclient==0.8.0
nbconvert==7.6.0
nbformat==5.9.0
nest-asyncio==1.5.6
networkx==3.2.1
notebook==6.5.4
notebook_shim==0.2.3
numba==0.60.0
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
omegaconf==2.3.0
onemetric==0.1.2
onnxruntime==1.18.1
onnxruntime-gpu==1.18.1
openai==0.28.0
opencv-python==4.8.1.78
opencv-python-headless==4.10.0.84
orjson==3.10.6
overrides==7.3.1
packaging==23.1
pandas==2.1.3
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
Pillow==10.1.0
platformdirs==3.8.0
pooch==1.8.2
prometheus-client==0.17.0
protobuf==5.27.2
psutil==5.9.8
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyarrow==16.1.0
pycparser==2.21
pydantic==2.8.2
pydantic_core==2.20.1
pydeck==0.9.1
pydub==0.25.1
Pygments==2.15.1
PyMatting==1.1.12
pyparsing==3.1.2
pytesseract==0.3.10
python-dateutil==2.8.2
python-dotenv==1.0.1
python-json-logger==2.0.7
python-multipart==0.0.9
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.0
referencing==0.31.0
regex==2024.5.15
rembg==2.0.57
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rpds-py==0.13.1
ruff==0.5.1
safetensors==0.4.3
scikit-image==0.24.0
scipy==1.11.4
seaborn==0.13.0
semantic-version==2.10.0
Send2Trash==1.8.2
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.4.1
stack-data==0.6.2
starlette==0.37.2
streamlit==1.36.0
supervision==0.1.0
sympy==1.12
tenacity==8.5.0
tensorboardX==2.6.2.2
terminado==0.17.1
thop==0.1.1.post2209072238
tifffile==2024.7.2
tinycss2==1.2.1
tokenizers==0.19.1
toml==0.10.2
tomlkit==0.12.0
toolz==0.12.1
torch==2.1.1
torchvision==0.16.1
tornado==6.3.2
tqdm==4.66.1
traitlets==5.9.0
transformers==4.42.3
triton==2.1.0
typer==0.12.3
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
ujson==5.10.0
ultralytics==8.1.19
urllib3==2.1.0
uvicorn==0.30.1
uvloop==0.19.0
validators==0.22.0
watchdog==4.0.1
watchfiles==0.22.0
wcwidth==0.2.6
webencodings==0.5.1
websocket-client==1.6.1
websockets==11.0.3
yarl==1.9.4
zipp==3.17.0
